OCPBUGS-43692: Fix LB Type Defaulting with ProviderParameters Set
In the case that `alreadyAdmitted` was true and the IngressController's
spec had `providerParameters` set, `setDefaultProviderParameters`
incorrectly forced the LB type to `Classic`.

This fix removes `alreadyAdmitted` as a parameter from
`setDefaultProviderParameters` and completely prevents it from being
invoked when `alreadyAdmitted`, preventing the update of defaulted
values after an IngressController has been admitted.

Previously, `setDefaultProviderParameters` always ensured that
`specLB.ProviderParameters.AWS` was initialized. Now that the function
is conditionally invoked, an additional nil check has been added to
handle cases when it's not called.

An unit test has been added for this scenario to prevent regression.